demo_k8_jobs_unzip_gcs/requirements.txt
google-cloud-storage
zipfile
io